third_party/blink/web_tests/platform/mac/external/wpt/svg/animations/slider-switch-expected.txt

This is a testharness.js-based test.
FAIL Check correct event bases for onclick assert_equals: expected "block" but got "none"
Harness: the test ran to completion.
